requirements.txt
python-wordpress-xmlrpc==2.3
termcolor==1.1.0
wheel==0.24.0
